clear up some compiler warnings (non-generics)
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1004082 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/src/java/org/apache/solr/schema/RandomSortField.java
@@ -167,7 +167,7 @@ public class RandomSortField extends FieldType {
 
           @Override
           public int intVal(int doc) {
-            return (int)hash(doc+seed);
+            return hash(doc+seed);
           }
 
           @Override
